Make integer narrowings explicit in the mule SNMP layer (-Wconversion)
Add explicit casts at the size_t/long to fixed-width narrowing sites flagged by
-Wconversion in the SNMP backend: ASN_INTEGER long->int32 value reads, the
int->float scale, errstat long->int, the key-passphrase length and the OID size
narrows. SNMP INTEGER is 32-bit per spec and the sizes are small, so the casts
are behavior-preserving and document intent.

+++ src/Oid.cpp
@@ -64,7 +64,7 @@ void Oid::assign ( std::string oidOfInterest )
 	}
 	m_oidVector.push_back(oidOfInterest.substr(previous, current - previous));
 
-	m_oidSize = m_oidVector.size();
+	m_oidSize = static_cast<uint32_t>(m_oidVector.size());
 
 }
